2010-04-25	Joel Sherrill <joel.sherrilL@OARcorp.com>
	* console/console.c, ide/pcmcia_ide.c, startup/bspstart.c,
	startup/imbx8xx.c: Fix some warnings.

+++ c/src/lib/libbsp/powerpc/mbx8xx/ide/pcmcia_ide.c
@@ -219,7 +219,7 @@ void mbx8xx_pcmciaide_read_block
 | Input Parameters:                                                         |
 \*-------------------------------------------------------------------------*/
  int minor,
- uint16_t   block_size,
+ uint32_t   block_size,
  rtems_blkdev_sg_buffer *bufs,
  uint32_t   *cbuf,
  uint32_t   *pos
@@ -276,7 +276,7 @@ void mbx8xx_pcmciaide_write_block
 | Input Parameters:                                                         |
 \*-------------------------------------------------------------------------*/
  int minor,
- uint16_t   block_size,
+ uint32_t   block_size,
  rtems_blkdev_sg_buffer *bufs,
  uint32_t   *cbuf,
  uint32_t   *pos
